+++ src/types/SubmitMonitor.ts
@@ -1,0 +1,4 @@
+export type SubmitMonitor = {
+  dataTimestamp: number;
+  txHash: string;
+};