+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - update umbrella SDK
 - update ethers to v5.7.2
+- use mapping (instead of file) to monitor last tx
 ## [7.7.1] - 2023-03-01
 ### Changed
 - use `StaticJsonRpcProvider` to reduce RPC calls